src/requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.1
Django==5.2.1
django-author==1.2.0
django-filter==25.1
django-picklefield==3.3
django-q2==1.8.0
django-safedelete==1.4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
flake8==7.2.0
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
psycopg2==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
setuptools-git==1.2
sqlparse==0.5.3
uritemplate==4.1.1
